input, fix for Debian Lenny Xlib changes
Fix delayed response to keyboard events.

Fixes a regression introduced by an upgrade of Debian GNU/Linux Lenny Xlib and
Xcb layers.  The symptom was that keyboard events would be delayed, they would
be buffered by Xlib before our code would handle them.  The cause was the use
of select(2) and the handling of the X event queue by the client.

The change is to allow Xlib a chance to read the X socket before asking Xlib
for the events.  Then, all events are processed, before proceeding.