full screen fixes
Now restores original screen resolution after normal play, and after
disconnection by the server in outfit screen.

Lessons learned:

1.  an XF86VidModeSwitchToMode call needs to be flushed to the X server before
the program exits, otherwise the call will be ignored (run down of the process
does not flush the library or socket buffers),

2.  an atexit call cannot be used with X11, since if it is activated during an
X error the X library has a lock and further X calls (in the function given to
atexit) cause a failure to lock ... and are therefore ineffective.