metaserver window, persistent
Makes the metaserver window persistent, in that a clicking on a server
launches a new window for playing, using a UNIX forked process with a
new display connection.

(By comparison, BRMH 2.4 implements this as a fork and exec of itself with new
command line arguments containing the server name.  However, that method does
not take into account any command line arguments given to the first process.)

The change to scroll_thumb_gc and scroll_thumb_pixmap allow the x11window.c
layer to be called again.  There was no reason to free these resources.