rework login event loop and state machine, fast guest
Reimplementation of getname(), using a more detailed state machine, so that
screen refresh events are fully handled, and it is made clear to the user
when we are waiting for the server to respond.

Also added a fast guest login feature for the metaserver list, a middle
mouse button click on a server will attempt a login as guest and join the
most under-represented team, complete with feedback to the metaserver list
in case the server refuses guest login.

Also tested automatic login properly, and verified that it works, in all
cases except where the password is null.

Added a brief pointer to maintainer on the README login screen.