team selection realignment, disconnection during play
When a disconnection occurs during play, the old behaviour was to hang, the
revised behaviour was to exit, the new behaviour is to continue to redraw
the tactical and galactic, report the loss of connection to warning window,
and allow q for quit.

Add an explanation of the message of the day, and the auto quit.