include key file on target system
By packaging for distributions from source in an automatic fashion, we
lose the public half of the key that used to be sent in to the metaserver
administration when we did the builds ourselves.

This change places a copy of the public key in the package so that if we
ever have a need for it the user can provide it.