network resilience, report stall of data stream
Previously, if the data stream from the server stalled, for example due
to a network cable being pulled, modem hanging up, internet routing
loss, or crash of the server operating system, the client would hang.

Now, the client notices the stall in the data stream after 1.1 seconds
of no packets received.  It reports it in the warning window.  User is
expected to either wait for restoration or close the client window.

Sudden termination of the server process remains handled by the lost
connection logic.  This patch deals with loss of streaming in situations
where the connection is not explicitly broken.