SP_RANK, allow st_rank to escape limits
Review following feedback:

2) Socket.c, handlePlyrLogin, rejecting packets with rank >= nranks.
This wasn't an issue when NUMRANKS was static, but once the  # of
ranks is an unknown, this check fails due to the fact that player
login packets are sent before the server tells the client the value
for nranks via the individual rank packets.  For Netrek XP, I
commented out this check and instead moved sanity checks on rank to
playerlist and newwin.