review and correct torp explosion animation
Originally, with five updates per second, with NUMDETFRAMES as five, a
torp explosion would animate over one second using those five frames.

With the addition of ten updates per second for starbases, torps began
to be shown as half a second animations of those same five frames.  With
the widening of ten updates per second support to non-starbases, people
got used to half a second torp explosions.

With our move to 25 or 50 updates per second, torps got shown as an even
shorter time but the same five frames.  So we tried to fix it.

This change restores the one second animation, and as a result all the
frames of the bitmap explosion is shown.