fix cloak pixmap animation
Animation using pixmaps of the cloak (the gradual loss of image pixels
over the progress of the cloaking) was not working correctly due to the
ten updates per second fix from many years ago ... specifically the
timing was too short; ships vanished far too quickly, and on decloak
there was one animation frame using bitmap graphics instead of the
pixmaps.

The former was caused by failure to count updates correctly;
CLOAK_PHASES was adjusted in local.c only and not in x11sprite.c.

The latter was caused by choosing the sprite composition *before*
updating p_cloakphase.  Moving the call to S_Ship fixes that.