Limit det to server frame rate
Previously a det was limited to 10 Hz, which was the server frame rate,
so that a server would not receive on average more than one det packet
per simulation frame.  Not clear why though, but the comment is clear,
and it dated from when servers were 10 Hz and maximum update rate was
the same.

Change mstime() utility function to ustime(), and cascade.

Limit det to server frame rate instead of 10 Hz.