Fix invalid read of size 4
When testing with Valgrind, by starting the client to show the server
list and then pressing the q key to exit, an invalid read of size 4
was reported in the call to W_IsMapped made by metainput.

The window had been deleted by metadone in response to the q key, but
the event loop continued.

Exit the event loop when an event handler returns true.